implement get_next_event(), and do parsing.
Updated main to call systemd_analyze::get_next_event(), added code to
parse the output from systemd-analyze, and transform it into a
DateTime<Local>.  This DateTime<Local> gets passed back to main.

This returned DateTime<Local> is the 'target' time.
In main we have an infinite loop, that gets the current time, and checks
it against the target time to see if the current time is the same or
greater than the current time.

It does this by getting a TimeDelta by subtracting dt_now from target.
If this TimeDelta's num_seconds() method returns zero or a negative
number, we fire the event.  In the case of this spike, we return, and
exit the program immediately.

This code will be used in the worker thread of walls, for the slideshow
mode.